Code cleanup and fixes
Cleaned up unnecessary code when finding the initial current_node.
Also, improved it to handle values of infinity correctly (to not add a
predecessor that is infinity away just because there’s nothing else
better). Also, improved handling for when the elements of index_values
are not the output of range(), and encapsulated the example code.